[QuickStart] Update session_id to have 64 bits
Test: existing unittests pass
Fixed: b/300110064
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4860363
Commit-Queue: Brittany Hartmire <bhartmire@google.com>
Cr-Commit-Position: refs/heads/main@{#1196706}

+++ chrome/browser/ash/login/oobe_quick_start/connectivity/connection.cc
@@ -103,7 +103,7 @@ void Connection::Close(
 }
 
 void Connection::RequestWifiCredentials(
-    int32_t session_id,
+    int64_t session_id,
     RequestWifiCredentialsCallback callback) {
   // Build the Wifi Credential Request payload
   SessionContext::SharedSecret secondary_shared_secret =
@@ -121,7 +121,7 @@ void Connection::RequestWifiCredentials(
                              std::move(on_response_received));
 }
 
-void Connection::NotifySourceOfUpdate(int32_t session_id,
+void Connection::NotifySourceOfUpdate(int64_t session_id,
                                       NotifySourceOfUpdateCallback callback) {
   if (base::CommandLine::ForCurrentProcess()->HasSwitch(
           kQuickStartTestForcedUpdateSwitch)) {
